% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendChat.R
\name{sendChat}
\alias{sendChat}
\title{DingTalk Robot send message}
\usage{
sendChat(webhook = NULL, msgtype = "text", content = NULL,
  title = NULL, text = NULL, messageUrl = NULL, picUrl = NULL,
  singleTitle = NULL, singleURL = NULL, btns = NULL,
  btnOrientation = "0", hideAvatar = "0", links = NULL,
  atMobiles = NULL, isAtAll = FALSE, result = FALSE)
}
\arguments{
\item{webhook}{Webhook applied on Dingtalk}

\item{msgtype}{Message type, includes text, markdown, link, actionCard and  feedCard. Default text.}

\item{content}{Message content}

\item{title}{Message title}

\item{text}{Message text}

\item{messageUrl}{Message URL}

\item{picUrl}{Picture URL}

\item{singleTitle}{Title of the single button. when singleTitle and singleURL are set, btns option is not available.}

\item{singleURL}{The URL of the singleTitle button}

\item{btns}{Button information list. Including title(button title), actionURL(the URL triggered by clicking the button). when singleTitle and singleURL are set, btns option is not available.}

\item{btnOrientation}{btns arranged direction, Vertical or horizontal. 0-vertically, 1-horizontally.}

\item{hideAvatar}{The sender's profile picture, 0-show, 1-hide}

\item{links}{The links list. Each link includes title, messageURL and picURL}

\item{atMobiles}{The mobile phone number of the person}

\item{isAtAll}{All the people, default is FALSE}

\item{result}{Return message result, default FALSE.}
}
\description{
DingTalk Robot send message
}
